git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@7218 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ TacticalAI/Movement.cpp
@@ -450,7 +450,8 @@ INT32 InternalGoAsFarAsPossibleTowards(SOLDIERTYPE *pSoldier, INT32 sDesGrid, IN
 	INT32 sOrigin;
 	UINT16 usMaxDist;
 	UINT8 ubDirection,ubDirsLeft,ubDirChecked[8],fFound = FALSE;
-	INT8 bAPsLeft, fPathFlags;
+	// anv: changed to INT16 since Soldier->bActionPoints are INT16 too
+	INT16 bAPsLeft, fPathFlags;
 	//DBrot: More Rooms
 	//UINT8 ubRoomRequired = 0, ubTempRoom;
 	UINT16 usRoomRequired = 0, usTempRoom;
